Add IssueServiceExtensions.NewIssue() and tests
The way to create a new issue in a Jira instance is using the Issue class’s constructor, but this requires the root Jira object. I am purposefully avoiding having the root Jira object available via dependency injection; hence this extension method. I might have designed the SDK so that the Issue class does not depend on the root Jira class and all issue operations use the services, but ¯\_(ツ)_/¯